Fix prototype of strtoll_l, strtoull_l, towlower_l and towupper_l

+++ 9/sources/android/support/include/stdlib.h
@@ -44,8 +44,8 @@ void _Exit(int);
 int                  mbtowc(wchar_t *pwc, const char *pmb, size_t max);
 long                 strtol_l(const char *nptr, char **endptr, int base, locale_t loc);
 unsigned long        strtoul_l(const char *nptr, char **endptr, int base, locale_t loc);
-long long            strtoll_l(const char *nptr, char **endptr, size_t base, locale_t loc);
-unsigned long long   strtoull_l(const char *nptr, char **endptr, size_t base, locale_t loc);
+long long            strtoll_l(const char *nptr, char **endptr, int base, locale_t loc);
+unsigned long long   strtoull_l(const char *nptr, char **endptr, int base, locale_t loc);
 long double          strtold_l(const char *nptr, char **endptr, locale_t loc);
 #endif // !__LP64__
